requirements.txt
tensorflow-cpu
numpy
opencv-python-headless
scikit-learn
matplotlib
requests
tqdm
Flask
